mail templates and util services

+++ src/main/java/com/m99/userloginsystem/dao/user/UserPersonalDetailsDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.m99.userloginsystem.entity.user.UserPersonalDetails;
 
-public interface UserPersonalDetailsDao extends JpaRepository<UserPersonalDetails, Integer>{
+public interface UserPersonalDetailsDao extends JpaRepository<UserPersonalDetails, Long>{
 
 }
