+++ src/main/java/com/m99/userloginsystem/dao/user/UserSecurityDetailsDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.m99.userloginsystem.entity.user.UserSecurityDetails;
 
-public interface UserSecurityDetailsDao extends JpaRepository<UserSecurityDetails, Integer>{
+public interface UserSecurityDetailsDao extends JpaRepository<UserSecurityDetails, Long>{
 
 }
